refactor: Rename admin-aside.tsx to auth-aside.tsx and remove user-aside.tsx

+++ src/actions/address/delete-user-address.ts
@@ -1,17 +1,24 @@
 'use server'
 
+import { auth } from '@/auth.config'
 import { prisma } from '@/lib'
 import { revalidatePath } from 'next/cache'
 
 export const deleteUserAddress = async (id: string) => {
   try {
+    const session = await auth()
+    if (!session) return { ok: false, message: 'No autorizado' }
+
     await prisma.userAddress.delete({
       where: {
         id
       }
     })
+
     revalidatePath('/shop/address')
     revalidatePath('/dashboard/addresses')
+    revalidatePath('/admin/users/[id]')
+
     return {
       ok: true,
       message: 'Dirección eliminada correctamente'
